Create Presentation of World_Happiness_2015_2020.pptx
This project involves making a structurally and visually cohesive presentation slides in Microsoft Powerpoint 2021 based on the "World Happiness Score Comparison, 2015 - 2020" data analytics project I completed in Tableu. This project is the estimation of a real presentation I would have made if given this particular case to analyze and present. Data visuals on the slides mostly come from the data analysis and visualization project on the same topic I completed in Tableu. The speaker notes on most slides contain thorough background information on the content of each slide, which I would not have read through in a real-life presentation.
For the interactive data visualization on the "World Happiness Score Comparison, 2015 - 2020", please visit my vizz on Tableu Public website:
https://public.tableau.com/views/World_Happiness_2015_2020/Dashboard?:language=en-US&:sid=&:display_count=n&:origin=viz_share_link